Refactor method of patching
Now uses mixin classes and wraps loggers instead of modifying them. This
means that getting the same logger using `bracelogger.get_logger` and
`logging.getLogger` and using different formats for them is now
possible.

It also means that `LogRecord` objects no longer have a `getMessage`
attribute on them which would get confused with other extra attributes
since that function is on the `BracelogRecordMixin` instead of the
object.